// src/interfaces/response.ts
export interface ICreatorResponse {
    id: string;
    avatar: string;
    name: string;
    default_user: string;
    career: string;
    instagram: string;
    tiktok: string;
    twitter: string;
    youtube: string;
    github: string;
    linkedin: string;
    facebook: string;
    twitch: string;
    discord: string;
    website: string;
}